backend/requirements.txt
# ── Core FastAPI stack ────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ── Django + Channels (Dynatrace/WebSocket side) ─────────────────
django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
channels==4.1.0
channels-redis==4.2.0

# ── LDAP ─────────────────────────────────────────────────────────
ldap3==2.9.1

# ── JWT ──────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0

# ── Settings / validation ────────────────────────────────────────
pydantic-settings==2.2.1
python-decouple==3.8

# ── Redis ────────────────────────────────────────────────────────
redis==5.0.4
hiredis==2.3.2

# ── HTTP ─────────────────────────────────────────────────────────
httpx==0.27.0
